split out reaction.go
